rename function validator module and constructor
